requirements.txt
requests
python-dotenv
flask
sqlalchemy
shodan
colorama
rich
